+++ core/data/src/main/java/ru/shawarma/core/data/datasources/MainOrderRemoteDataSource.kt
@@ -35,7 +35,7 @@ class MainOrderRemoteDataSource @Inject constructor(
         }
 
 
-    override suspend fun getOrder(token: String, id: Int): Result<OrderResponse> =
+    override suspend fun getOrder(token: String, id: Long): Result<OrderResponse> =
         safeServiceCall(dispatcher){
             orderService.getOrder(token, id)
         }
