+++ core/data/src/main/java/ru/shawarma/core/data/datasources/OrderRemoteDataSource.kt
@@ -10,7 +10,7 @@ interface OrderRemoteDataSource {
 
     suspend fun getOrders(token: String, offset: Int, count: Int): Result<List<OrderResponse>>
 
-    suspend fun getOrder(token: String, id: Int): Result<OrderResponse>
+    suspend fun getOrder(token: String, id: Long): Result<OrderResponse>
 
     suspend fun createOrder(token: String, request: CreateOrderRequest): Result<OrderResponse>
