+++ core/data/src/main/java/ru/shawarma/core/data/entities/OrderMenuItemRequest.kt
@@ -1,6 +1,6 @@
 package ru.shawarma.core.data.entities
 
 data class OrderMenuItemRequest(
-    val menuItemId: Int,
+    val menuItemId: Long,
     val amount: Int
 )
