+++ feature/settings/src/main/java/ru/shawarma/settings/NavigationCommand.kt
@@ -1,5 +1,5 @@
 package ru.shawarma.settings
 
 sealed interface NavigationCommand{
-    data class ToOrderModule(val orderId: Int): NavigationCommand
+    data class ToOrderModule(val orderId: Long): NavigationCommand
 }
